Homework-4 clean
