+++ homework-4/src/main/java/ru/otus/homework/service/MessageService.java
@@ -1,24 +1,8 @@
 package ru.otus.homework.service;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.MessageSource;
-import org.springframework.stereotype.Service;
+public interface MessageService {
 
-import java.util.Locale;
+    String getMessage(String code);
 
-@Service
-@RequiredArgsConstructor
-public class MessageService implements MessageServiceImpl {
-
-    private final MessageSource messageSource;
-
-    @Override
-    public String getMessage(String code) {
-        return messageSource.getMessage(code, null, Locale.getDefault());
-    }
-
-    @Override
-    public String getMessage(String code, Object[] args) {
-        return messageSource.getMessage(code, args, Locale.getDefault());
-    }
+    String getMessage(String code, Object[] args);
 }
